feat(nap): removed the params in the request in favor of a body field
the body param is the way we handle url queries and stuff now. Instead
of having a body and a params variable, we just use body like this:

body = { { encoded = true, data = "here is some data" } }

encoded is optional and will default to true for a GET request
and false for a POST request.

encoded = true:  --data-urlencoded
encoded = false:  --data

this consolidates the data in the request to make it less confusing than
having a "body" and a "params" field and makes it more like curl. since
curl only uses the --data flags for both the body of a request and the
url queries